Hide visibility of the browse button
Why this change?

Because users will sometimes select both: (i) folders and (ii) files at
the same time. When the user clicks "select" in the file explorer
according to the operating system they are using, nothing happens. And
there is no warning/explanation given to the user. To prevent the
confusion which results, I am forcing the readers to drag and drop ONLY,
till a solution to the folder/file selection problem is found.